% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{find_intemidates}
\alias{find_intemidates}
\title{find_intemidates}
\usage{
find_intemidates(
  peak_table_unlabel,
  peak_table_label,
  sample_info = NULL,
  path = ".",
  polarity = c("positive", "negative"),
  control_group = c("WT_UA", "WT_13CUA"),
  case_group = c("ygeX_UA", "ygeX_13CUA"),
  mz_tol = 10,
  rt_tol = 0.05,
  p_value_cutoff = 0.05,
  p_adjust = TRUE,
  fold_change_cutoff = 20,
  is_recognize_adducts = TRUE
)
}
\arguments{
\item{peak_table_unlabel}{unlabeled peak table}

\item{peak_table_label}{labeled peak table}

\item{path}{working directory. Default: "."}

\item{control_group}{label of control groups. e.g. c("WT_UA", "WT_13CUA")}

\item{case_group}{label of case groups. e.g. c('ygeX_UA', 'ygeX_13CUA')}

\item{mz_tol}{Default: 10 ppm}

\item{rt_tol}{Default: 0.05 min}

\item{p_value_cutoff}{Default: 0.05}

\item{fold_change_cutoff}{Default: 20}
}
\description{
find_intemidates
}
\examples{
\dontrun{
find_intemidates(peak_table_unlabel = '230830_Csp_YgeX_WT_UA.xlsx',
                 peak_table_label = '230830_Csp_YgeX_WT_13CUA.xlsx',
                 path = '~/Project/00_Uric_Acid_project/20230831_isotope_tracing/',
                 control_group = c("WT_UA", "WT_13CUA"),
                 case_group = c('ygeX_UA', 'ygeX_13CUA'),
                 mz_tol = 10,
                 rt_tol = 0.05,
                 p_value_cutoff = 0.05,
                 fold_change_cutoff = 20)
}
}
\author{
Zhiwei Zhou
}
